CUSTOMER_SUPPORT_AGENT/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
asyncer==0.0.7
attrs==25.3.0
auth0-python==4.10.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chainlit==2.5.5
charset-normalizer==3.4.2
chevron==0.14.0
Chroma==0.2.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.126.0
crewai-tools==0.0.1
cryptography==45.0.4
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
durationpy==0.10
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflection==0.5.1
instructor==1.8.3
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
json5==0.12.0
json_repair==0.46.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-openai==0.3.22
langchain-text-splitters==0.3.8
langsmith==0.3.45
Lazify==0.4.0
litellm==1.68.0
literalai==0.1.201
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.9.3
mdurl==0.1.2
mem0ai==0.1.107
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.0
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.75.0
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.40.9
opentelemetry-instrumentation-anthropic==0.40.9
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-bedrock==0.40.9
opentelemetry-instrumentation-chromadb==0.40.9
opentelemetry-instrumentation-cohere==0.40.9
opentelemetry-instrumentation-crewai==0.40.9
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-google-generativeai==0.40.9
opentelemetry-instrumentation-groq==0.40.9
opentelemetry-instrumentation-haystack==0.40.9
opentelemetry-instrumentation-lancedb==0.40.9
opentelemetry-instrumentation-langchain==0.40.9
opentelemetry-instrumentation-llamaindex==0.40.9
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.40.9
opentelemetry-instrumentation-mcp==0.40.9
opentelemetry-instrumentation-milvus==0.40.9
opentelemetry-instrumentation-mistralai==0.40.9
opentelemetry-instrumentation-ollama==0.40.9
opentelemetry-instrumentation-openai==0.40.9
opentelemetry-instrumentation-pinecone==0.40.9
opentelemetry-instrumentation-qdrant==0.40.9
opentelemetry-instrumentation-replicate==0.40.9
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.40.9
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.40.9
opentelemetry-instrumentation-transformers==0.40.9
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.40.9
opentelemetry-instrumentation-watsonx==0.40.9
opentelemetry-instrumentation-weaviate==0.40.9
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.52b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.2.1
portalocker==2.10.1
posthog==3.25.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.14.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.41.3
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
traceloop-sdk==0.40.9
traitlets==5.14.3
typer==0.16.0
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uptrace==1.31.0
urllib3==2.4.0
uv==0.7.12
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0